+++ app/src/main/kotlin/me/xizzhu/android/joshua/annotated/bookmarks/list/BookmarkItem.kt
@@ -16,11 +16,8 @@
 
 package me.xizzhu.android.joshua.annotated.bookmarks.list
 
-import android.graphics.Typeface
 import android.text.Spannable
 import android.text.SpannableStringBuilder
-import android.text.style.RelativeSizeSpan
-import android.text.style.StyleSpan
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.TextView
@@ -28,6 +25,8 @@ import me.xizzhu.android.joshua.R
 import me.xizzhu.android.joshua.core.Constants
 import me.xizzhu.android.joshua.core.Settings
 import me.xizzhu.android.joshua.core.VerseIndex
+import me.xizzhu.android.joshua.ui.createBookNameSizeSpan
+import me.xizzhu.android.joshua.ui.createBookNameStyleSpan
 import me.xizzhu.android.joshua.ui.recyclerview.BaseItem
 import me.xizzhu.android.joshua.ui.recyclerview.BaseViewHolder
 import me.xizzhu.android.joshua.ui.updateSettingsWithPrimaryText
@@ -38,8 +37,8 @@ data class BookmarkItem(val verseIndex: VerseIndex, private val bookName: String
                         val onClick: (VerseIndex) -> Unit)
     : BaseItem(R.layout.item_bookmarks, { inflater, parent -> BookmarkItemViewHolder(inflater, parent) }) {
     companion object {
-        private val BOOK_NAME_STYLE_SPAN = StyleSpan(Typeface.BOLD)
-        private val BOOK_NAME_SIZE_SPAN = RelativeSizeSpan(0.9F)
+        private val BOOK_NAME_SIZE_SPAN = createBookNameSizeSpan()
+        private val BOOK_NAME_STYLE_SPAN = createBookNameStyleSpan()
         private val SPANNABLE_STRING_BUILDER = SpannableStringBuilder()
     }