+++ CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 - Should wait for 2.5 seconds before tracking reading progress
 - Should not track reading progress when no translation is available
 - Should not fail when a parallel translation is not available
+- Updated dependencies:
+  - Grardle to 5.6
 
 #### v0.9.0 (2019-08-10)
 - Improved UI for reading